requirements.txt
Appium-Python-Client==0.24